latest version

+++ deploy/script/user-rpc-dev.sh
@@ -1,15 +1,22 @@
-# !/bin/bash
+#!/bin/bash
 reso_addr="crpi-4ll7w1qk44676hc7.cn-hangzhou.personal.cr.aliyuncs.com/go-chat-im/user-rpc-dev"
 tag="latest"
+container_name="go-chat-im-user-rpc-dev"
 
-container_name = "go-chat-im-user-rpc-dev"
+username="Peninsular"
+password="yujie2018"
 
-docker stop ${container_name}
+echo $password | docker login $reso_addr -u $username --password-stdin
 
-docker rm ${container_name}
-
-docker rmi ${reso_addr}:${tag}
-
+# 停止容器
+docker stop ${container_name} || true
+# 删除容器
+docker rm ${container_name} || true
+# 检查镜像是否存在，如果存在则删除
+if docker images | grep -q "${reso_addr}:${tag}"; then
+  docker rmi ${reso_addr}:${tag}
+fi
+# 拉取最新的镜像
 docker pull ${reso_addr}:${tag}
-
-docker run -d --name ${container_name} -p 10000:10000 ${reso_addr}:${tag}
+# 运行新的容器
+docker run -d --name ${container_name} -p 10000:10000${reso_addr}:${tag}
